Move duplicate get/delete code up into FileSystemStateManager
